🧪 Steward: Harden test assertions and improve categorization
- Loosen brittle HTML-string assertions in BladeShellRenderingTest and AdminListAccessibilityTest to focus on functional attributes and text.
- Relocate MeetingPhotoMigrationServiceTest from Unit to Integration as it hits the database and filesystem.
- Use RefreshDatabase and randomized, slugged identifiers in MeetingPhotoMigrationServiceTest for better parallel isolation and database constraint compliance.
